Properly separate providers from the rest of port context.
