+++ crates/opte-api/src/lib.rs
@@ -53,7 +53,7 @@ pub use ulp::*;
 ///
 /// We rely on CI and the check-api-version.sh script to verify that
 /// this number is incremented anytime the oxide-api code changes.
-pub const API_VERSION: u64 = 37;
+pub const API_VERSION: u64 = 38;
 
 /// Major version of the OPTE package.
 pub const MAJOR_VERSION: u64 = 0;
